Fix clipping of page content.
The operator 'n' has to come after the 'W' or 'W*': 'W', 'W*' leave the
current path also in place. If we continue to make a new path, it will
also include all of the old one. So, 'n' is used to clear the path *after*
'W'/'W*'. The clip path is prepared, and we are ready to put the page
content inside it.

File test/bugs/Kalender.pdf is one hit by this bug.

Many thanks to Johannes Brödel for reporting this bug and the example file.